feat: Add server id to replicated writes
Adds the server id to replicated writes. Changes type from u8 to u16 for more potential write servers. Also updates the server struct in Delorean to no longer have locks. Locking will happen higher up, likely with the server being arc_swap'd out on any configuration update.

+++ delorean_data_types/src/data.rs
@@ -121,7 +121,7 @@ impl fmt::Display for ReplicatedWrite {
 }
 
 pub fn lines_to_replicated_write(
-    writer: u8,
+    writer: u16,
     sequence: u64,
     lines: &[ParsedLine<'_>],
     rules: &DatabaseRules,
